+++ src/main/java/com/example/thymeleaf/repository/LightBulbsRoomIntRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LightBulbsRoomIntRepository extends JpaRepository<LightBulbsRoomInt,Integer> {
+public interface LightBulbsRoomIntRepository extends JpaRepository<LightBulbsRoomInt,Long> {
 }
